Consistent `use features` and "no experimental warnings".
`state` is great, but no code in this distribution needs it, so no need to
mention it. Add no warnings prederef for to avoid warnings on v5.20.0.
Do this in all modules even where the feature(s) are not used as consistency
seems a bit clearer than minimal precision. (And not doing this leaves a trap
for anyone editing the code in the future and only testing on a more recent
version of perl)