Paranoid error handling with clear reporting in _authenticate
*We*, not user code, are handling errors from the external API here - it's
much easier to debug "the impossible" if we report it clearly, rather than
explode with Moose constraint violations. Given that we now have much more
detailed error reporting (both location and probable cause) it seems better
to use croak() rather than confess() - this lets the caller choose whether
they want to show/log a stack backtrace, rather than use forcing it on them.

Both success and error responses *should* be JSON, so we test for that
before deciding whether we are handling 2xx status codes or 4xx/5xx.

The original code was assuming the v1/v2 API error response format.

This still seems to be used in the API used to get an access token, but the
v3 API TrueLayer has changed to the JSON format described in RFC 7807.
(The migration guide doesn't mention this.) Handle both formats.