Refactor api_post() and api_get() to use _ua_request()
This eliminates _process_response() and inlines _process_response_new() into
_ua_request(). It changes the format of error messages slightly, so adapt
the tests. Remove the TODO - GET failures now report 'GET' not 'POST'.

api_post() and api_get() now take a parameter to explicitly state whether
the API is expected to return JSON, with the default being true. This
replaces the previous code which would heuristically treat an empty body
(or, '0') as "return nothing" and always return the value otherwise.

It's unclear whether we need this for either GET or POST - of all the
APIs I could see, only those using DELETE could return an empty body.
(With a 200 code on success, not a 204, which feels wrong.)